@@ -17,8 +17,8 @@
 		<maven.compiler.source>${java.version}</maven.compiler.source>
 		<maven.compiler.target>${java.version}</maven.compiler.target>
 
-		<spring-framework.version>5.3.1</spring-framework.version>
-		<spring-boot.version>2.4.0</spring-boot.version>
+		<spring-framework.version>5.3.13</spring-framework.version>
+		<spring-boot.version>2.6.0</spring-boot.version>
 	</properties>
 
 	<dependencies>
@@ -26,7 +26,7 @@
 		<dependency>
 			<groupId>org.dbunit</groupId>
 			<artifactId>dbunit</artifactId>
-			<version>2.7.0</version>
+			<version>2.7.2</version>
 			<scope>provided</scope>
 		</dependency>
 		<!-- Extended: DbUnit for Spring -->
@@ -47,12 +47,12 @@
 		<dependency>
 			<groupId>commons-io</groupId>
 			<artifactId>commons-io</artifactId>
-			<version>2.8.0</version>
+			<version>2.11.0</version>
 		</dependency>
 		<dependency>
 			<groupId>com.github.jsqlparser</groupId>
 			<artifactId>jsqlparser</artifactId>
-			<version>3.2</version>
+			<version>4.2</version>
 		</dependency>
 
 		<!-- Tests -->
@@ -71,7 +71,7 @@
 		<dependency>
 			<groupId>org.hsqldb</groupId>
 			<artifactId>hsqldb</artifactId>
-			<version>2.5.1</version>
+			<version>2.6.1</version>
 			<scope>test</scope>
 		</dependency>
 	</dependencies>
@@ -81,7 +81,7 @@
 			<plugin>
 				<groupId>org.jacoco</groupId>
 				<artifactId>jacoco-maven-plugin</artifactId>
-				<version>0.8.6</version>
+				<version>0.8.7</version>
 				<executions>
 					<execution>
 						<goals>
